Moved run_metrics  to cpp (cherry picked from commit f9ed1d2)
