+++ src/interop/logic/plot/plot_flowcell_map.cpp
@@ -8,6 +8,7 @@
 #include "interop/logic/plot/plot_flowcell_map.h"
 
 #include "interop/logic/metric/metric_value.h"
+#include "interop/logic/metric/q_metric.h"
 
 namespace illumina { namespace interop { namespace logic { namespace plot {
