+++ src/interop/logic/plot/plot_qscore_heatmap.cpp
@@ -8,6 +8,7 @@
 #include "interop/logic/plot/plot_qscore_heatmap.h"
 
 #include "interop/model/plot/bar_point.h"
+#include "interop/logic/metric/q_metric.h"
 
 namespace illumina { namespace interop { namespace logic { namespace plot
 {
@@ -177,6 +178,24 @@ namespace illumina { namespace interop { namespace logic { namespace plot
         if(metrics.run_info().flowcell().surface_count()>1 && options.is_specific_surface())
             title += " " + options.surface_description();
         data.set_title(title);
+    }
+    /** Count number of rows for the heat map
+     *
+     * @param metrics run metrics
+     * @return number of rows
+     */
+    size_t count_rows_for_heatmap(const model::metrics::run_metrics& metrics)
+    {
+        return metrics.get_set< model::metrics::q_metric >().max_cycle();
+    }
+    /** Count number of columns for the heat map
+     *
+     * @param metrics run metrics
+     * @return number of columns
+     */
+    size_t count_columns_for_heatmap(const model::metrics::run_metrics& metrics)
+    {
+        return logic::metric::max_qval(metrics.get_set< model::metrics::q_metric >());
     }
